first unit tests for app.py
